removed more unneeded things
