+++ VAMobile/src/components/index.ts
@@ -89,8 +89,6 @@ export { default as TextLineWithIcon } from './TextLineWithIcon'
 export * from './TextLineWithIcon'
 export { default as InlineTextWithIcons } from './InlineTextWithIcons'
 export * from './InlineTextWithIcons'
-export { default as SignoutButton } from './SignoutButton'
-export * from './SignoutButton'
 export { default as PhotoAdd } from './PhotoAdd'
 export * from './PhotoAdd'
 export { default as PhotoPreview } from './PhotoPreview'
